+++ geonode/base/models.py
@@ -83,10 +83,6 @@ from geonode.notifications_helper import (
     send_notification,
     get_notification_recipients)
 from geonode.people.enumerations import ROLE_VALUES
-from geonode.base.thumb_utils import (
-    thumb_path,
-    thumb_size,
-    remove_thumbs)
 
 from pyproj import transform, Proj
 
@@ -1656,6 +1652,11 @@ class ResourceBase(PolymorphicModel, PermissionLevelMixin, ItemBase):
     # Note - you should probably broadcast layer#post_save() events to ensure
     # that indexing (or other listeners) are notified
     def save_thumbnail(self, filename, image):
+        from geonode.thumbs.utils import (
+            thumb_path,
+            thumb_size,
+            remove_thumbs
+        )
         upload_path = thumb_path(filename)
         try:
             # Check that the image is valid
