+++ geonode/base/utils.py
@@ -36,7 +36,7 @@ from guardian.shortcuts import get_perms, remove_perm, assign_perm
 
 from geonode.layers.models import Dataset
 from geonode.base.models import ResourceBase, Link, Configuration
-from geonode.base.thumb_utils import (
+from geonode.thumbs.utils import (
     get_thumbs,
     remove_thumb)
 from geonode.utils import get_legend_url
